fix the drawer flash issue
